requirements.txt
python-telegram-bot
schedule
qrcode
python-dotenv
colorlog